Moving to a new tree model (snapshot)
