+++ AngouriMath/AngouriMath.csproj
@@ -37,8 +37,4 @@
     <PackageReference Include="System.Memory" Version="4.5.4" />
     <PackageReference Include="Nullable" Version="1.2.1" PrivateAssets="all" />
   </ItemGroup>
-  
-  <ItemGroup>
-    <Folder Include="Functions\Continuous\" />
-  </ItemGroup>
 </Project>
