+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/AnalyticalSolver.cs
