+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/CommonDenominatorSolver.cs
