+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/FractionedPolynoms.cs
