+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/PolynomialSolver.cs
