+++ AngouriMath/Functions/Continuous/Algebra/AnalyticalSolving/TrigonometricSolver.cs
