+++ AngouriMath/Functions/Continuous/Algebra/EquationSolver.cs
