+++ AngouriMath/Functions/Continuous/Algebra/Limits/Limit.cs
