+++ AngouriMath/Functions/Continuous/Algebra/NumericalSolving/NewtonSolver.cs
