+++ AngouriMath/Functions/Continuous/Algebra/NumericalSolving/NumericalInequalitySolver.cs
