+++ AngouriMath/Functions/Continuous/Evaluation/Patterns.cs
