+++ AngouriMath/Functions/Continuous/Evaluation/Simplificator.cs
