+++ AngouriMath/Functions/Continuous/Evaluation/TreeAnalyzer/Factorization.cs
