+++ AngouriMath/Functions/Continuous/Evaluation/TreeAnalyzer/Sort.cs
