+++ AngouriMath/Functions/Continuous/Evaluation/TrigonometricTableValues.cs
